Extract cidr.lang dependent code in //cpp to optional runtime dep
Trivial move
