+++ cpp/src/com/google/idea/blaze/cpp/oclang/BlazeCppAutoImportHelper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.idea.blaze.cpp;
+package com.google.idea.blaze.cpp.oclang;
 
 import com.google.common.collect.ImmutableList;
 import com.google.idea.blaze.base.settings.Blaze;
