+++ cpp/src/com/google/idea/blaze/cpp/oclang/CppSyncStatusContributor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.idea.blaze.cpp.syncstatus;
+package com.google.idea.blaze.cpp.oclang;
 
 import com.google.idea.blaze.base.model.BlazeProjectData;
 import com.google.idea.blaze.base.model.primitives.LanguageClass;
